requirements.txt
httpx>=0.27,<1
pycryptodome>=3.20,<4
diskcache>=5.6,<6
python-dotenv>=1.0,<2
google-genai>=1.0,<2
